@@ -141,7 +141,7 @@ class _LocationSheetBaseState extends State<LocationSheetBase> {
     return ListTile(
       leading: Icon(_selectedIcon.icon),
       title: Text(_selectedIcon.title),
-      tileColor: Colors.blue.withOpacity(0.1),
+      tileColor: Colors.blue.withValues(alpha: 0.1),
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8),
       ),
@@ -224,7 +224,7 @@ class IconSelector extends StatelessWidget {
     return ListTile(
       leading: Icon(selectedIcon.icon),
       title: Text(selectedIcon.title),
-      tileColor: Colors.blue.withOpacity(0.1),
+      tileColor: Colors.blue.withValues(alpha: 0.1),
       shape: RoundedRectangleBorder(
         borderRadius: BorderRadius.circular(8),
       ),
