@@ -1,8 +1,15 @@
 import 'package:flutter/material.dart';
+import 'package:flutter/foundation.dart';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
+import 'package:map_app/theme.dart';
+import 'package:map_app/utils.dart';
+import 'package:map_app/widgets/auth/google_oauth_screen.dart';
+import 'package:map_app/widgets/auth/login_success.dart';
 import 'package:map_app/widgets/map/main_map.dart';
+
 import 'data/datastore/factory.dart';
 import 'data/state/providers/initialize_tiles_provider.dart';
+import 'data/auth/auth_providers.dart';
 
 void main() async {
   WidgetsFlutterBinding.ensureInitialized();
@@ -20,15 +27,59 @@ class MyApp extends ConsumerWidget {
 
   @override
   Widget build(BuildContext context, WidgetRef ref) {
+    // Initialize tiles
     ref.watch(initializeTilesProvider);
+
+    if (kDebugMode) {
+      print("Building MyApp");
+    }
+    TextTheme textTheme = createTextTheme(context, "Roboto", "Libre Baskerville");
+    MaterialTheme theme = MaterialTheme(textTheme);
 
     return MaterialApp(
       title: 'Turbo',
-      theme: ThemeData(
-        colorScheme: ColorScheme.fromSeed(seedColor: const Color.fromRGBO(0, 95, 126, 100)),
-        useMaterial3: true,
-      ),
-      home: const MapControllerPage(),
+      debugShowCheckedModeBanner: false,
+      theme:  theme.light(),
+      // Define routes for navigation
+      routes: {
+        '/': (context) => const HomeWrapper(),
+        '/login/callback': (context) => const GoogleAuthCallbackPage(),
+        '/login/success': (context) => const LoginSuccessPage(),
+      },
+      initialRoute: '/',
     );
+  }
+}
+
+class HomeWrapper extends ConsumerStatefulWidget {
+  const HomeWrapper({super.key});
+
+  @override
+  ConsumerState<HomeWrapper> createState() => _HomeWrapperState();
+}
+
+class _HomeWrapperState extends ConsumerState<HomeWrapper> {
+
+  @override
+  void initState() {
+    super.initState();
+
+    // Initialize auth silently without blocking the app
+    _initializeAuth();
+  }
+
+  Future<void> _initializeAuth() async {
+    try {
+      await ref.read(authStateProvider.notifier).initialize();
+    } catch (e) {
+      if (kDebugMode) {
+        print('Auth initialization error: $e');
+      }
+    }
+  }
+
+  @override
+  Widget build(BuildContext context) {
+    return const MapControllerPage();
   }
 }
