@@ -22,14 +22,26 @@ class LocationMarkers extends ConsumerWidget {
       data: (locations) => MarkerLayer(
         markers: locations.map((location) {
           final namedIcon = iconService.getIcon(location.icon);
+          // Define marker scale
+          const double markerScale = 1.0;
+          // Calculate the offset to adjust marker position
+          const double verticalOffset = 30.0 * markerScale; // Half of the marker height
+
           return Marker(
             width: 80.0,
-            height: 80.0,
+            height: 90.0,
             point: location.position,
-            child: MapIcon(
-              namedIcon: namedIcon,
-              title: location.title,
-              onTap: () => _showEditSheet(context, ref, location),
+            alignment: Alignment.bottomCenter,
+            // Offset the marker up by half its height so the point is at the exact location
+            rotate: false,
+            child: Transform.translate(
+              offset: const Offset(0, -verticalOffset),
+              child: MapMarker(
+                namedIcon: namedIcon,
+                title: location.title,
+                onTap: () => _showEditSheet(context, ref, location),
+                scale: markerScale,
+              ),
             ),
           );
         }).toList(),
@@ -52,87 +64,194 @@ class LocationMarkers extends ConsumerWidget {
   }
 }
 
-class MapIcon extends StatefulWidget {
+class MapMarker extends StatefulWidget {
   final NamedIcon namedIcon;
   final String title;
   final VoidCallback onTap;
+  final double scale;
 
-  const MapIcon({
+  const MapMarker({
     super.key,
     required this.namedIcon,
     required this.title,
     required this.onTap,
+    this.scale = 1.0,
   });
 
   @override
-  State<MapIcon> createState() => _MapIconState();
+  State<MapMarker> createState() => _MapMarkerState();
 }
 
-class _MapIconState extends State<MapIcon> {
+class _MapMarkerState extends State<MapMarker> {
   bool _isHovering = false;
   bool _isPressed = false;
 
+  static const double baseWidth = 38.0;
+  static const double baseHeight = 50.0;
+  static const double iconSize = 22.0;
+  static const double iconPadding = 2.0;
+
   @override
   Widget build(BuildContext context) {
-    return MouseRegion(
-      cursor: SystemMouseCursors.click,
-      onEnter: (_) => setState(() => _isHovering = true),
-      onExit: (_) => setState(() => _isHovering = false),
-      child: GestureDetector(
-        onTapDown: (_) => setState(() => _isPressed = true),
-        onTapUp: (_) => setState(() => _isPressed = false),
-        onTapCancel: () => setState(() => _isPressed = false),
-        onTap: widget.onTap,
-        child: Column(
-          children: [
-            AnimatedContainer(
-              duration: const Duration(milliseconds: 100),
-              width: 40, // Adjust size as needed
-              height: 40, // Adjust size as needed
-              decoration: BoxDecoration(
-                color: _isHovering || _isPressed ? Colors.blue.shade100 : Colors.white,
-                shape: BoxShape.circle,
-                boxShadow: [
-                  BoxShadow(
-                    color: Colors.black.withOpacity(0.2),
-                    blurRadius: 4,
-                    offset: const Offset(0, 2),
+    final colorScheme = Theme.of(context).colorScheme;
+    final textTheme = Theme.of(context).textTheme;
+
+    // Calculate scaled dimensions
+    final width = baseWidth * widget.scale;
+    final height = baseHeight * widget.scale;
+    final iconSizeScaled = iconSize * widget.scale;
+    final iconPaddingScaled = iconPadding * widget.scale;
+
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: [
+        // Marker with droplet shape - fixed position regardless of hover state
+        GestureDetector(
+          onTapDown: (_) => setState(() => _isPressed = true),
+          onTapUp: (_) => setState(() => _isPressed = false),
+          onTapCancel: () => setState(() => _isPressed = false),
+          onTap: widget.onTap,
+          child: MouseRegion(
+            cursor: SystemMouseCursors.click,
+            onEnter: (_) => setState(() => _isHovering = true),
+            onExit: (_) => setState(() => _isHovering = false),
+            child: Stack(
+              clipBehavior: Clip.none,
+              children: [
+                // Droplet shape
+                CustomPaint(
+                  size: Size(width, height),
+                  painter: DropletPainter(
+                    color: colorScheme.surface, // Surface color for the base
+                    isPressed: _isPressed,
+                    isHovered: _isHovering,
+                    outlineColor: colorScheme.outline.withValues(alpha: 0.5),
+                    scale: widget.scale,
                   ),
-                ],
-                border: Border.all(
-                  color: _isHovering || _isPressed ? Colors.blue : Colors.transparent,
-                  width: 2,
                 ),
-              ),
-              child: Center(
-                child: Icon(
-                  widget.namedIcon.icon,
-                  color: _isHovering || _isPressed ? Colors.blue.shade700 : Colors.blue,
-                  size: 24, // Adjust size as needed
+
+                // Icon positioned inside the top circular part of the droplet
+                Positioned(
+                  top: iconPaddingScaled,
+                  left: iconPaddingScaled,
+                  child: Container(
+                    width: width - (iconPaddingScaled * 2),
+                    height: width - (iconPaddingScaled * 2),
+                    decoration: BoxDecoration(
+                      color: colorScheme.surface, // Surface color for the icon background
+                      shape: BoxShape.circle,
+                    ),
+                    alignment: Alignment.center,
+                    child: Icon(
+                      widget.namedIcon.icon,
+                      color: colorScheme.primary, // Primary color for the icon itself
+                      size: iconSizeScaled,
+                    ),
+                  ),
                 ),
-              ),
+              ],
             ),
-            const SizedBox(height: 4), // Space between icon and text
-            Container(
-              padding: const EdgeInsets.symmetric(horizontal: 4, vertical: 2),
-              decoration: BoxDecoration(
-                color: Colors.white.withOpacity(0.8),
-                borderRadius: BorderRadius.circular(4),
-              ),
-              child: Text(
-                widget.title,
-                style: TextStyle(
-                  fontSize: 12,
-                  color: Colors.blue.shade700 ,
-                  fontWeight: FontWeight.bold,
-                ),
-                maxLines: 1,
-                overflow: TextOverflow.ellipsis,
-              ),
-            ),
-          ],
+          ),
         ),
-      ),
+
+        // Label always appears below the marker, visibility controlled by state
+        AnimatedOpacity(
+          opacity: (_isHovering || _isPressed) ? 1.0 : 0.0,
+          duration: const Duration(milliseconds: 200),
+          child: Container(
+            margin: EdgeInsets.only(top: 4.0 * widget.scale),
+            padding: EdgeInsets.symmetric(
+              horizontal: 6.0 * widget.scale,
+              vertical: 2.0 * widget.scale,
+            ),
+            decoration: BoxDecoration(
+              color: colorScheme.surfaceContainerHighest,
+              borderRadius: BorderRadius.circular(4.0 * widget.scale),
+              boxShadow: [
+                BoxShadow(
+                  color: Colors.black.withValues(alpha: 0.1),
+                  blurRadius: 2.0 * widget.scale,
+                  offset: Offset(0, 1.0 * widget.scale),
+                ),
+              ],
+            ),
+            child: Text(
+              widget.title,
+              style: textTheme.labelSmall?.copyWith(
+                fontWeight: FontWeight.w500,
+                color: colorScheme.onSurface,
+                fontSize: (textTheme.labelSmall?.fontSize ?? 10) * widget.scale,
+              ),
+              maxLines: 1,
+              overflow: TextOverflow.ellipsis,
+            ),
+          ),
+        ),
+      ],
     );
+  }
+}
+
+// Custom painter to draw the droplet shape
+class DropletPainter extends CustomPainter {
+  final Color color;
+  final Color outlineColor;
+  final bool isPressed;
+  final bool isHovered;
+  final double scale;
+
+  DropletPainter({
+    required this.color,
+    required this.outlineColor,
+    this.isPressed = false,
+    this.isHovered = false,
+    this.scale = 1.0,
+  });
+
+  @override
+  void paint(Canvas canvas, Size size) {
+    final double width = size.width;
+    final double height = size.height;
+    final double circleRadius = width / 2;
+
+    final paintFill = Paint()
+      ..color = color
+      ..style = PaintingStyle.fill;
+
+    final paintStroke = Paint()
+      ..color = outlineColor
+      ..style = PaintingStyle.stroke
+      ..strokeWidth = (isPressed || isHovered ? 1.0 : 0.5) * scale;
+
+    final path = Path();
+
+    // Start at the bottom point
+    path.moveTo(width / 2, height);
+
+    // Draw left curve
+    path.quadraticBezierTo(0, height * 0.6, 0, height * 0.4);
+
+    // Draw top-left curve of the circle
+    path.quadraticBezierTo(0, 0, circleRadius, 0);
+
+    // Draw top-right curve of the circle
+    path.quadraticBezierTo(width, 0, width, height * 0.4);
+
+    // Draw right curve
+    path.quadraticBezierTo(width, height * 0.6, width / 2, height);
+
+    // Draw the fill and stroke
+    canvas.drawShadow(path, Colors.black.withValues(alpha: 0.2), 2.0 * scale, true);
+    canvas.drawPath(path, paintFill);
+    canvas.drawPath(path, paintStroke);
+  }
+
+  @override
+  bool shouldRepaint(covariant DropletPainter oldDelegate) {
+    return oldDelegate.color != color ||
+        oldDelegate.outlineColor != outlineColor ||
+        oldDelegate.isPressed != isPressed ||
+        oldDelegate.isHovered != isHovered ||
+        oldDelegate.scale != scale;
   }
 }
