@@ -7,6 +7,9 @@ import '../../data/model/marker.dart';
 import '../../data/state/providers/location_provider.dart';
 import 'location_base_sheet.dart';
 
+// Import the shared button components
+import '../auth/primary_button.dart';
+
 class CreateLocationSheet extends ConsumerStatefulWidget {
   final Marker? location;
   final LatLng? newLocation;
@@ -22,6 +25,7 @@ class CreateLocationSheetState extends ConsumerState<CreateLocationSheet> {
   late NamedIcon _selectedIcon;
   late TextEditingController _nameController;
   late TextEditingController _descriptionController;
+  bool _isLoading = false;
 
   @override
   void initState() {
@@ -40,19 +44,59 @@ class CreateLocationSheetState extends ConsumerState<CreateLocationSheet> {
 
   @override
   Widget build(BuildContext context) {
-    return Padding(
+    final colorScheme = Theme.of(context).colorScheme;
+    final textTheme = Theme.of(context).textTheme;
+
+    return Container(
+      decoration: BoxDecoration(
+        color: colorScheme.surface,
+        borderRadius: const BorderRadius.vertical(top: Radius.circular(28)),
+      ),
       padding: EdgeInsets.only(
-        bottom: MediaQuery.of(context).viewInsets.bottom,
-        left: 16,
-        right: 16,
-        top: 16,
+        bottom: MediaQuery.of(context).viewInsets.bottom + 24,
+        left: 24,
+        right: 24,
+        top: 12,
       ),
       child: Form(
         key: _formKey,
         child: Column(
           mainAxisSize: MainAxisSize.min,
           children: [
-            _buildHeader(),
+            // Drag handle
+            Center(
+              child: Container(
+                width: 32,
+                height: 4,
+                decoration: BoxDecoration(
+                  color: colorScheme.outlineVariant,
+                  borderRadius: BorderRadius.circular(2),
+                ),
+              ),
+            ),
+            const SizedBox(height: 16),
+            // Header
+            Row(
+              mainAxisAlignment: MainAxisAlignment.spaceBetween,
+              children: [
+                Text(
+                  'New Marker',
+                  style: textTheme.titleLarge?.copyWith(
+                    fontWeight: FontWeight.bold,
+                  ),
+                ),
+                IconButton(
+                  onPressed: () => Navigator.pop(context),
+                  icon: const Icon(Icons.close),
+                  style: IconButton.styleFrom(
+                    foregroundColor: colorScheme.onSurfaceVariant,
+                    backgroundColor: colorScheme.surfaceContainerHighest.withValues(alpha: 0.3),
+                  ),
+                ),
+              ],
+            ),
+            const SizedBox(height: 24),
+            // Form fields
             LocationFormFields(
               nameController: _nameController,
               descriptionController: _descriptionController,
@@ -60,91 +104,80 @@ class CreateLocationSheetState extends ConsumerState<CreateLocationSheet> {
               onIconSelected: (icon) => setState(() => _selectedIcon = icon),
             ),
             const SizedBox(height: 32),
-            _buildSaveButton(),
-            const SizedBox(height: 16),
+            // Save button
+            SizedBox(
+              width: double.infinity,
+              child: PrimaryButton(
+                text: 'Save Marker',
+                onPressed: _saveLocation,
+                isLoading: _isLoading,
+              ),
+            ),
+            const SizedBox(height: 8),
           ],
         ),
       ),
     );
   }
 
-  Widget _buildHeader() {
-    return Padding(
-      padding: const EdgeInsets.only(bottom: 16.0),
-      child: Row(
-        mainAxisAlignment: MainAxisAlignment.spaceBetween,
-        children: [
-          Text(
-            'Ny markering',
-            style: TextStyle(fontWeight: FontWeight.bold, color: Colors.grey[600]),
-          ),
-          IconButton(
-            onPressed: () => Navigator.pop(context),
-            icon: const Icon(Icons.close),
-          )
-        ],
-      ),
-    );
-  }
-
-  Widget _buildSaveButton() {
-    return SizedBox(
-      width: 128,
-      child: ElevatedButton.icon(
-        icon: const Icon(Icons.save, color: Colors.white),
-        label: const Text('Lagre', style: TextStyle(color: Colors.white)),
-        onPressed: () => _saveLocation(context),
-        style: ElevatedButton.styleFrom(
-          backgroundColor: Colors.blueGrey,
-          foregroundColor: Colors.white,
-          shape: RoundedRectangleBorder(
-            borderRadius: BorderRadius.circular(20),
-          ),
-          padding: const EdgeInsets.symmetric(vertical: 16, horizontal: 12),
-        ),
-      ),
-    );
-  }
-
-  Future<void> _saveLocation(BuildContext context) async {
+  void _saveLocation() async {
     if (_formKey.currentState!.validate()) {
-      final locationProvider = ref.read(locationNotifierProvider.notifier);
-
-      final newMarker = Marker.fromMap({
-        'title': _nameController.text,
-        'description': _descriptionController.text,
-        'icon': _selectedIcon.title,
-        'latitude': widget.location?.position.latitude ?? widget.newLocation!.latitude,
-        'longitude': widget.location?.position.longitude ?? widget.newLocation!.longitude,
-      });
+      setState(() => _isLoading = true);
 
       try {
+        final locationProvider = ref.read(locationNotifierProvider.notifier);
+
+        final newMarker = Marker.fromMap({
+          'title': _nameController.text,
+          'description': _descriptionController.text,
+          'icon': _selectedIcon.title,
+          'latitude': widget.location?.position.latitude ?? widget.newLocation!.latitude,
+          'longitude': widget.location?.position.longitude ?? widget.newLocation!.longitude,
+        });
+
         await locationProvider.addLocation(newMarker);
 
-        if (!context.mounted) return;
-
+        if (!mounted) return;
         Navigator.of(context).pop(newMarker);
       } catch (e) {
-        // Handle the error
-        if (!context.mounted) return;
-
+        if (!mounted) return;
         Navigator.of(context).pop();
-
-        ScaffoldMessenger.of(context).showSnackBar(
-          SnackBar(
-            content: Text(_getErrorMessage(e)),
-            backgroundColor: Colors.red,
-          ),
-        );
+        _showErrorSnackBar(context, e);
+      } finally {
+        if (mounted) {
+          setState(() => _isLoading = false);
+        }
       }
     }
   }
 
+  void _showErrorSnackBar(BuildContext context, dynamic error) {
+    final colorScheme = Theme.of(context).colorScheme;
+
+    ScaffoldMessenger.of(context).showSnackBar(
+      SnackBar(
+        content: Row(
+          children: [
+            Icon(Icons.error_outline, color: colorScheme.onErrorContainer),
+            const SizedBox(width: 16),
+            Expanded(
+              child: Text(_getErrorMessage(error)),
+            ),
+          ],
+        ),
+        behavior: SnackBarBehavior.floating,
+        backgroundColor: colorScheme.errorContainer,
+        shape: RoundedRectangleBorder(borderRadius: BorderRadius.circular(12)),
+        margin: const EdgeInsets.all(16),
+      ),
+    );
+  }
+
   String _getErrorMessage(dynamic error) {
-  if (error is DatabaseException) {
+    if (error is DatabaseException) {
       return 'Database error. Please try again later.';
     } else {
-      return 'An unexpected error occurred. Please try again. ${error.toString()}';
+      return 'An unexpected error occurred. Please try again.';
     }
   }
 }
